Update VipNovel to new Madara Version (#178)
After BoxNovel, NovelTrench and LightNovelHeaven, VipNovel seems to also have updated to what I would not title as a new Madara version.
Therefore, it uses a Ajax request without form data in the post request, where the used URL contains the novel information. See #174 for more or less the same changes and more information on the Ajax URL.

Seems the default HTML object of the title changed from h3 to h1 with the update. Due to not wanting to affect the other Madara extensions for now, I just updated that selection for VipNovel.

Has been tested, chapters load again and the title after opening a novel is correct. I assume the same problem as with BoxNovel, NovelTrench and LightNovelHeaven will occur, i.e. that already bookmarked novels of VipNovel will get double chapters. This I did not test as I can not get chapters to load using the old method. I assume there will be the newly added chapters from the new Ajax usage and the old ones, which could have been read previously. Those old ones do not work. Due to what reason I am not sure as I do not have them. Maybe there was also an URL change of the chapters that results in them not getting overwritten.