Update Madara Extension Defaults, Change Foxaholic  (#190)
- Madara: Update default values to the ones that all extensions used but Foxaholic
- Extensions that use Madara: Remove values that are the same as default
- Madara: Update title identifier
- Madara: Remove badges from title (like HOT or NEW)
- Madara: Update chapter release identifier
- Madara: Add a h1 with chapter title before chapter text

These changes are specifically target for the the follow purposes:

    Foxaholic: Changed default values fixes Ajax usage (see 

    Update VipNovel to new Madara Version #178 )
    LightNovelBastion: Badges removed from title
    LightNovelHeaven: Title fixed
    Extensions: Some extensions have a proper title in chapter text, most do not. Add a proper title before every chapter.

The added title is more of a preference on my side. I dislike opening a chapter and having to do extra button presses to check what the chapter title is. Disadvantage: The extensions already having one will have two, first the h1, that got added, followed by typically a h3, but not necessarily. It is sometimes inconsistent among chapters of the same novel.

Changes should not break #175 .