Add Order to Madara, Update WuxiaDotBlog (#220)
This is #208 without the libs and WWVolare changes. Therefore, closes #207 .
As imo the point of libraries is missed by auto installing them and therefore there seems to be the desire to not add unnecessary libraries and instead duplicate code between extensions, I removed the libraries and moved one needed function from a removed lib to WuxiaDotBlog.
The WWVolare got scrapped.

Added the Order of NovelChapter to Madara and WuxiaBlog.
Fixing default listing ended up in partial rewrite of WuxiaDotBlog until the error of it starting with page 0 instead of 1 was noticed.
Changed WuxiaDotBlog to HTML chapters.
Shrank URL in WuxiaDotBlog where applicable.